# docs/04-api/Enterprise_AI_API_Gateway_OpenAPI_3.1_V1.2_Final.yaml
openapi: 3.1.0
info:
  title: Enterprise AI & API Gateway
  version: 1.2.0
  summary: Final executable HTTP contract for Enterprise AI & API Gateway V1
  description: Derived from PRD V1.4, HLD V1.1 corrections, ERD V1.2 Final, Physical DB/Alembic V1.2 Final, API Specification
    V1.2.2 Final, frozen M0-M9 execution contracts, and the M2.6 Normal API Registry Contract Reconciliation addendum.
    This artifact consolidates final V1 semantics for Normal API, OpenAI-compatible LLM, Admin browser sessions,
    traffic-admission health, idempotency, certification, credential invalidation, SSRF/TLS and SSE.
servers:
- url: https://gateway.example.com
  description: Production example
security: []
...
paths:
...
  /v1/models:
    get:
      tags:
      - LLM Data Plane
      operationId: listModels
      summary: List aliases/models invokable by the current API key
      security:
      - GatewayApiKey: []
      responses:
        '200':
          description: OpenAI-compatible model list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/admin/llm/providers:
    get:
      tags:
      - LLM Registry
      operationId: listLlmProviders
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProviderPage'
    post:
      tags:
      - LLM Registry
      operationId: createLlmProvider
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmProviderCreate'
      responses:
        '201':
          description: Provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProvider'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/targets:
    get:
      tags:
      - LLM Registry
      operationId: listLlmTargets
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Target list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmTargetPage'
    post:
      tags:
      - LLM Registry
      operationId: createLlmTarget
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmTargetCreate'
      responses:
        '201':
          description: Target created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmTarget'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/targets/{target_id}/credential:
    put:
      tags:
      - LLM Registry
      operationId: rotateLlmTargetCredential
      summary: Atomically rotate provider credential
      description: 'Idempotency-Key is mandatory.


        When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.'
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/TargetId'
      - $ref: '#/components/parameters/RequiredIdempotencyKey'
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderCredentialWrite'
      responses:
        '200':
          description: Credential metadata only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialMetadata'
        '409': *id001
  /api/v1/admin/llm/models:
    get:
      tags:
      - LLM Registry
      operationId: listLlmModels
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmModelPage'
    post:
      tags:
      - LLM Registry
      operationId: createLlmModel
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmModelCreate'
      responses:
        '201':
          description: Model created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmModel'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/models/{model_id}/capabilities:
    put:
      tags:
      - LLM Registry
      operationId: replaceModelCapabilities
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/ModelId'
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - capabilities
              additionalProperties: false
              properties:
                capabilities:
                  type: array
                  uniqueItems: true
                  items:
                    $ref: '#/components/schemas/ModelCapability'
      responses:
        '200':
          description: Capability set replaced
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/aliases:
    get:
      tags:
      - LLM Registry
      operationId: listLlmAliases
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Alias list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmAliasPage'
    post:
      tags:
      - LLM Registry
      operationId: createLlmAlias
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmAliasCreate'
      responses:
        '201':
          description: Alias created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmAlias'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/aliases/{alias_id}/targets:
    put:
      tags:
      - LLM Registry
      operationId: replaceAliasTargets
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/AliasId'
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - targets
              additionalProperties: false
              properties:
                targets:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/AliasTargetWrite'
      responses:
        '200':
          description: Ordered target set replaced atomically
        '409':
          $ref: '#/components/responses/ResourceConflict'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
...
  /api/v1/admin/llm/models/{model_id}/prices:
    parameters:
    - $ref: '#/components/parameters/ModelId'
    get:
      tags:
      - Pricing
      operationId: listModelPrices
      security:
      - AdminToken: []
      responses:
        '200':
          description: Price windows
    post:
      tags:
      - Pricing
      operationId: createModelPrice
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelPriceCreate'
      responses:
        '201':
          description: Price window created
        '409':
          $ref: '#/components/responses/PriceWindowConflict'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
...
  /api/v1/admin/llm/providers/{provider_id}:
    parameters:
    - $ref: '#/components/parameters/ProviderId'
    get:
      tags:
      - LLM Providers
      summary: Get LLM provider
      operationId: getLlmProvider
      security:
      - AdminToken: []
      responses:
        '200':
          description: Provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - LLM Providers
      summary: Update or disable LLM provider
      operationId: updateLlmProvider
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmProviderPatch'
      responses:
        '200':
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProvider'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/models/{model_id}:
    parameters:
    - $ref: '#/components/parameters/ModelId'
    get:
      tags:
      - LLM Models
      summary: Get LLM model
      operationId: getLlmModel
      security:
      - AdminToken: []
      responses:
        '200':
          description: Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmModel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - LLM Models
      summary: Update LLM model status or mutable metadata
      operationId: updateLlmModel
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmModelPatch'
      responses:
        '200':
          description: Updated model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmModel'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/llm/aliases/{alias_id}:
    parameters:
    - $ref: '#/components/parameters/AliasId'
    get:
      tags:
      - LLM Aliases
      summary: Get LLM alias
      operationId: getLlmAlias
      security:
      - AdminToken: []
      responses:
        '200':
          description: Alias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmAlias'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - LLM Aliases
      summary: Update or disable LLM alias
      operationId: updateLlmAlias
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmAliasPatch'
      responses:
        '200':
          description: Updated alias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmAlias'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
...
  /api/v1/admin/llm/targets/{target_id}:
    parameters:
    - $ref: '#/components/parameters/TargetId'
    get:
      tags:
      - LLM Registry
      operationId: getLlmTarget
      security:
      - AdminToken: []
      responses:
        '200':
          description: Provider target metadata including certification and Experimental override; no credential plaintext.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmTarget'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - LLM Registry
      operationId: updateLlmTarget
      security:
      - AdminToken: []
      description: Updates mutable target fields. When session-authenticated, X-CSRF-Token is mandatory. Certification result
        fields and credentials cannot be arbitrarily set here.
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmTargetPatch'
      responses:
        '200':
          description: Updated target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmTarget'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
...
  /api/v1/admin/model-prices/{price_id}:
    parameters:
    - $ref: '#/components/parameters/PriceId'
    patch:
      tags:
      - Pricing
      operationId: updateModelPrice
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: Adjust future/open price window subject to no-overlap exclusion constraint. Session-authenticated mutation
        requires X-CSRF-Token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelPricePatch'
      responses:
        '200':
          description: Updated price window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPrice'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PriceWindowConflict'
components:
  securitySchemes:
    GatewayApiKey:
      type: http
      scheme: bearer
      bearerFormat: gw_<secret>
      description: Data Plane credential only.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: adm_<secret>
      description: Per-admin Control Plane credential only.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: gw_admin_session
      description: Opaque Redis-backed Admin browser session cookie. Fail-closed when session store is unavailable.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque server-generated cursor
      schema:
        type: string
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required only when the operation is authenticated with AdminSessionCookie and the HTTP method is state-changing.
        Not required for direct Bearer adm_ authentication.
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
  responses:
    Unauthorized:
      $ref: '#/components/responses/_Unauthorized'
...
    RateLimited:
      description: Gateway or provider rate limit
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 0
        X-RateLimit-Limit:
          schema:
            type: integer
            minimum: 0
        X-RateLimit-Remaining:
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ModelList:
      type: object
      required:
      - object
      - data
      properties:
        object:
          type: string
          const: list
        data:
          type: array
          items:
            type: object
            required:
            - id
            - object
            - owned_by
            properties:
              id:
                type: string
              object:
                type: string
                const: model
              owned_by:
                type: string
                const: gateway
...
    LlmProviderCreate:
      type: object
      required:
      - name
      - provider_type
      additionalProperties: false
      properties:
        name:
          type: string
        provider_type:
          type: string
          enum:
          - OPENAI
          - ANTHROPIC
          - VLLM
          - GENERIC_OPENAI_COMPAT
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
          default: ACTIVE
    LlmProvider:
      allOf:
      - $ref: '#/components/schemas/LlmProviderCreate'
      - type: object
        properties:
          id:
            type: string
            format: uuid
...
    LlmTarget:
      allOf:
      - $ref: '#/components/schemas/LlmTargetCreate'
      - type: object
        required:
        - id
        - certification_status
        - allow_uncertified_runtime
        properties:
          id:
            type: string
            format: uuid
          certification_status:
            type: string
            enum:
            - UNVERIFIED
            - CERTIFIED
            - FAILED
          certified_at:
            type:
            - string
            - 'null'
            format: date-time
          certification_metadata:
            type:
            - object
            - 'null'
            additionalProperties: true
            description: Safe certification metadata/reference only; never credential/provider raw payload.
...
    LlmProviderPage:
      type: object
      required:
      - data
      - next_cursor
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LlmProvider'
        next_cursor:
          type:
          - string
          - 'null'
    LlmTargetPage:
      type: object
      required:
      - data
      - next_cursor
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LlmTarget'
        next_cursor:
          type:
          - string
          - 'null'
...